opencca: enable CCA-bound code execution on non-CCA Arm RK3588 SoC
This introduces support for FEAT_RME on non-CCA-capable
RK3588 SoC (Armv8.2) to emulate Arm CCA functionality on real hardware.

We introduce a new build option: ENABLE_OPENCCA=1 for RK3588 platform
that allows TFA to simulate a realm world within the normal world.

In conjunction with opencca TF-RMM, this enables to run CCA
realm VMs on real hardware -- in the normal world and for
research benchmark/evaluations and prototyping with real peripherals.

Build with:

	$(MAKE) -j$(NPROC) \
		PLAT=rk3588 \
		ENABLE_OPENCCA=1 \
		ENABLE_RME=1 \
		DEBUG=$(DEBUG) \
		LOG_LEVEL=$(LOG) \
		ENABLE_PAUTH=0 \
		ENABLE_FEAT_DIT=0 \
		RME_GPT_MAX_BLOCK=0

Evaluated on Radxa Rock 5b RK3588 16 GB RAM model.